requirements.txt
python-telegram-bot[job-queue]==22.5
python-dotenv>=1.0.1
tzdata>=2024.1; platform_system=="Windows"